Add breeds helpers
