+++ robustness/tools/imagenet_helpers.py
@@ -1,523 +1,354 @@
-import sys, os
-import json
+import os
 import numpy as np
-from collections import OrderedDict, Counter
-import operator
-import networkx as nx
-from ..datasets import DATASETS
+import json
+from itertools import product
 
-class ClassHierarchy():
+class Node():
     '''
-    Class representing a general ImageNet-style hierarchy.
+    Class for representing a node in the ImageNet/WordNet hierarchy. 
     '''
-    def __init__(self, info_dir):
+    def __init__(self, wnid, parent_wnid=None, name=""):
         """
         Args:
-            info_dir (str) : Path to hierarchy information files. Contains a 
-                "class_hierarchy.txt" file with one edge per line, a
-                "node_names.txt" mapping nodes to names, and "class_info.json".
+            wnid (str) : WordNet ID for synset represented by node
+            parent_wnid (str) : WordNet ID for synset of node's parent
+            name (str) : word/human-interpretable description of synset 
         """
-        
-        with open(f'{info_dir}/class_info.json', 'r') as f:
-            class_info = json.load(f)
+
+        self.wnid = wnid
+        self.name = name
+        self.class_num = -1
+        self.parent_wnid = parent_wnid
+        self.descendant_count_in = 0
+        self.descendants_all = set()
+    
+    def add_child(self, child):
+        """
+        Add child to given node.
+
+        Args:
+            child (Node) : Node object for child
+        """
+        child.parent_wnid = self.wnid
+    
+    def __str__(self):
+        return f'Name: ({self.name}), ImageNet Class: ({self.class_num}), Descendants: ({self.descendant_count_in})'
+    
+    def __repr__(self):
+        return f'Name: ({self.name}), ImageNet Class: ({self.class_num}), Descendants: ({self.descendant_count_in})'
+
+class ImageNetHierarchy():
+    '''
+    Class for representing ImageNet/WordNet hierarchy. 
+    '''
+    def __init__(self, ds_path, ds_info_path):
+        """
+        Args:
+            ds_path (str) : Path to ImageNet dataset
+            ds_info_path (str) : Path to supplementary files for the ImageNet dataset 
+                                 ('wordnet.is_a.txt', 'words.txt' and 'imagenet_class_index.json')
+                                 which can be obtained from http://image-net.org/download-API.
+
+        """
+        self.tree = {}
+
+        ret = self.load_imagenet_info(ds_path, ds_info_path)
+        self.in_wnids, self.wnid_to_name, self.wnid_to_num, self.num_to_name = ret
             
-        self.IN_WNIDS = [c['wnid'] for c in class_info]
-        self.WORDNET_TO_NUM = {k['wnid']: k['cid'] for k in class_info}
-
-        with open((f'{info_dir}/class_hierarchy.txt')) as f:
-            edges = [l.strip().split() for l in f.readlines()]
-            
-        self.graph = self._make_parent_graph(self.IN_WNIDS, edges)
-
-        with open((f'{info_dir}/node_names.txt')) as f:
-            mapping = [l.strip().split('\t') for l in f.readlines()]
-        self.NODE_NAME = {w[0]: w[1] for w in mapping}
-            
-        self.level_dict = self._make_level_dict(self.graph, root='n00001740')
-
-    @staticmethod
-    def _make_parent_graph(nodes, edges):
-        """
-        Create a graph for the predecessors of the nodes given.
-
-        Args:
-            nodes [str] : List of node names to traverse upwards.
-            edges [(str, str)] : Tuples of parent-child pairs.
-
-        Return:
-            networkx representation of the graph.
-        """
-
-        # create full graph
-        full_graph_dir = {}
-        for p, c in edges:
-            if p not in full_graph_dir:
-                full_graph_dir[p] = {c: 1}
-            else:
-                full_graph_dir[p].update({c: 1})
-                    
-        FG = nx.DiGraph(full_graph_dir)
-
-        # perform backward BFS to get the relevant graph
-        graph_dir = {}
-        todo = [n for n in nodes if n in FG.nodes()] # skip nodes not in graph
-        while todo:
-            curr = todo
-            todo = []
-            for w in curr:
-                for p in FG.predecessors(w):
-                    if p not in graph_dir:
-                        graph_dir[p] = {w: 1}
-                    else:
-                        graph_dir[p].update({w: 1})
-                    todo.append(p)
-            todo = set(todo)
-        
-        return nx.DiGraph(graph_dir)
-
-    @staticmethod
-    def _make_level_dict(graph, root):
-        """
-        Map nodes to their level within the hierarchy (top-down).
-
-        Args:
-            graph (networkx graph( : Graph representation of the hierarchy
-            root (str) : Hierarchy root.
-
-        Return:
-            Dictionary mapping node names to integer level.
-        """    
-
-        level_dict = {} 
-        todo = [(root, 0)] # (node, depth)
-        while todo:
-            curr = todo
-            todo = []
-            for n, d in curr:
-                if n not in level_dict:
-                    level_dict[n] = d
-                else:
-                    level_dict[n] = max(d, level_dict[n]) # keep longest path
-                for c in graph.successors(n):
-                    todo.append((c, d + 1))
-
-        return level_dict
-
-    def leaves_reachable(self, n):
-        """
-        Determine the leaves (ImageNet classes) reachable for a give node.
-
-        Args:
-            n (str) : WordNet ID of node
-
-        Returns:
-            leaves (list): List of WordNet IDs of the ImageNet descendants
-        """    
-        leaves = set()
-        todo = [n]
-        while todo:
-            curr = todo
-            todo = []
-            for w in curr:
-                for p in self.graph.successors(w):
-                    if p in self.IN_WNIDS:
-                        leaves.add(p)
-                    else:
-                        todo.append(p)
-            todo = set(todo)
-
-        # If the node itself is an ImageNet node
-        if n in self.IN_WNIDS: leaves = leaves.union([n])
-        return leaves
-
-    def node_name(self, n):
-        """
-        Name of a node.
-        """    
-        if n in self.NODE_NAME:
-            return self.NODE_NAME[n]
-        else:
-            return n
-
-    def print_node_info(self, nodes):
-        """
-        Prints basic information (name, number of ImageNet descendants) 
-        about a given set of nodes.
-
-        Args:
-            nodes (list) : List of WordNet IDs of relevant nodes
-        """    
-
-        for n in nodes:
-            if n in self.NODE_NAME:
-                print_str = f"{n}: {self.NODE_NAME[n]}"
-            else:
-                print_str = n
-
-            print_str += f" ({len(self.leaves_reachable(n))})"
-            print(print_str)
-
-    def traverse(self, nodes, direction='down', depth=20):
-        """
-        Find all nodes accessible from a set of nodes.
-
-        Args:
-            nodes (list) : List of WordNet IDs of relevant nodes
-            direction ("up"/"down"): Allowed traversal directions
-            depth (int): Maximum depth to traverse (from nodes)
-
-        Returns:
-            Set of nodes reachable within the desired depth, in the
-            desired direction.
-        """    
-
-        if not nodes or depth == 0:
-            return nodes
-
-        todo = []
-        for n in nodes:
-            if direction == 'down':
-                todo.extend(self.graph.successors(n))
-            else: 
-                todo.extend(self.graph.predecessors(n))
-        return nodes + self.traverse(todo, direction=direction, depth=depth-1)
-
-    def get_nodes_at_level(self, L, ancestor=None):
-        """
-        Find all superclasses at a specified depth.
-
-        Args:
-            L (int): Depth in hierarchy (from root node)
-            ancestor (str): (optional) WordNet ID that can be used to
-                            restrict the subtree in which superclasses
-                            are found
-
-        Returns:
-            nodes (list): Set of superclasses at that depth in 
-                                   the hierarchy
-        """            
-        if ancestor is not None:
-            valid = set(self.traverse([ancestor], direction="down"))
-
-        nodes= set()
-        for k, v in self.level_dict.items() :
-            if v == L and (ancestor is None or k in valid):
-                nodes.add(k)
+        with open(os.path.join(ds_info_path, 'wordnet.is_a.txt'), 'r') as f:
+            for line in f.readlines():
+                parent_wnid, child_wnid = line.strip('\n').split(' ')
+                parentNode = self.get_node(parent_wnid)
+                childNode = self.get_node(child_wnid)
+                parentNode.add_child(childNode)
                 
-        return nodes
-
-class BreedsDatasetGenerator():
-    '''
-    Class for generating datasets from ImageNet superclasses.
-    '''
-    def __init__(self, info_dir):    
-        self.hierarchy = ClassHierarchy(info_dir)
-
-    def split_superclass(self, superclass_wnid, Nsubclasses, balanced,
-                            split_type, rng=np.random.RandomState(2)):
-
-        """
-        Split superclass into two disjoint sets of subclasses.
-
-        Args:
-            superclass_wnid (str): WordNet ID of superclass node
-            Nsubclasses (int): Number of subclasses per superclass
-                               (not used when balanced is True)
-            balanced (bool): Whether or not the dataset should be
-                             balanced over superclasses
-            split_type ("good"/"bad"/"rand"): Whether the subclass
-                             partitioning should be more or less 
-                             adversarial or random
-            rng (RandomState): Random number generator
-
-        Returns:
-            class_ranges (tuple): Tuple of lists of subclasses
-        """ 
-
-        # Find a descendant of the superclass that has at least two children        
-        G = self.hierarchy.graph
-        node, desc = superclass_wnid, sorted(list(G.successors(superclass_wnid)))
-        while len(desc) == 1:
-            node = desc[0]
-            desc = sorted(list(G.successors(node)))
-        
-        # Map each descendant to its ImageNet subclasses
-        desc_map = {}
-        for d in desc:
-            dcurr = sorted(list(self.hierarchy.leaves_reachable(d)))
-            desc_map[d] = dcurr
-
-        # Map sorted by nodes that have the maximum number of children
-        desc_sorted = sorted(desc_map.items(), key=lambda x: -len(x[1]))
-
-        # If not balanced, we will pick as many subclasses as possible
-        # from this superclass (ignoring Nsubclasses)
-        if not balanced:
-            S = sum([len(d) for d in desc_map.values()])
-            assert S >= Nsubclasses
-            Nsubclasses = S
-            if Nsubclasses % 2 != 0:
-                Nsubclasses = max(Nsubclasses - 1, 2)
-
-        # Split superclasses into two disjoint sets
-        assert Nsubclasses % 2 == 0
-        Nh = Nsubclasses // 2
-
-        if split_type == "rand":
-            # If split is random, aggregate all subclasses, subsample
-            # the desired number, and then partition into two
-            desc_node_list = []
-            for d in desc_map.values():
-                desc_node_list.extend(d)
-            sel = rng.choice(sorted(desc_node_list), Nh * 2, replace=False)
-            split = (sel[:Nh], sel[Nh:])
-        else:
-            # If split is good, we will partition similar subclasses across
-            # both domains. If it is bad, similar subclasses will feature in
-            # one or the other
-
-            split, spare = ([], []), []
+        for wnid in self.in_wnids:
+            self.tree[wnid].descendant_count_in = 0
+            self.tree[wnid].class_num = self.wnid_to_num[wnid]
             
-            for k, v in desc_sorted:
-                l = [len(s) + 0 for s in split]
-                if split_type == "bad":            
-                    if l[0] <= l[1] and l[0] < Nh:
-                        if len(v) > Nh - l[0]: spare.extend(v[Nh-l[0]:])
-                        split[0].extend(v[:Nh-l[0]])
-                    elif l[1] < Nh:
-                        if len(v) > Nh - l[1]: spare.extend(v[Nh-l[1]:])
-                        split[1].extend(v[:Nh-l[1]])
-                else:
-                    if len(v) == 1:
-                        i1 = 1 if l[0] < Nh else 0
-                    else:
-                        i1 = min(len(v) // 2, Nh - l[0])
+        for wnid in self.in_wnids:
+            node = self.tree[wnid]
+            while node.parent_wnid is not None:
+                self.tree[node.parent_wnid].descendant_count_in += 1
+                self.tree[node.parent_wnid].descendants_all.update(node.descendants_all)
+                self.tree[node.parent_wnid].descendants_all.add(node.wnid)
+                node = self.tree[node.parent_wnid]
+        
+        del_nodes = [wnid for wnid in self.tree \
+                     if (self.tree[wnid].descendant_count_in == 0 and self.tree[wnid].class_num == -1)]
+        for d in del_nodes:
+            self.tree.pop(d, None)
                         
-                    if l[0] < Nh:
-                        split[0].extend(v[:i1])
-                    if l[1] < Nh:
-                        split[1].extend(v[i1:i1 + Nh-l[1]])
-            
-            if split_type == "bad":
-                l = [len(s) + 0 for s in split]
-                assert max(l) == Nh
-                if l[0] < Nh:
-                    split[0].extend(spare[:Nh - l[0]])
-                if l[1] < Nh:
-                    split[1].extend(spare[:Nh - l[1]])
-                
-        assert len(split[0]) == len(split[1]) and not set(split[0]).intersection(split[1])
-        class_ranges = ([self.hierarchy.WORDNET_TO_NUM[s] for s in split[0]],
-                        [self.hierarchy.WORDNET_TO_NUM[s] for s in split[1]])
-        
-        return class_ranges
+        assert all([k.descendant_count_in > 0 or k.class_num != -1 for k in self.tree.values()])
 
-    def get_superclasses(self, level, Nsubclasses=None,
-                         split=None, ancestor=None, balanced=True, 
-                         random_seed=2, verbose=False):
+        self.wnid_sorted = sorted(sorted([(k, v.descendant_count_in, len(v.descendants_all)) \
+                                        for k, v in self.tree.items()
+                                        ],
+                                        key=lambda x: x[2], 
+                                        reverse=True
+                                        ),
+                                key=lambda x: x[1], 
+                                reverse=True
+                                )
+
+    @staticmethod
+    def load_imagenet_info(ds_path, ds_info_path):
         """
-        Obtain a dataset composed of ImageNet superclasses with a desired
-        set of properties. 
-        (Optional) By specifying a split, one can parition the subclasses
-        into two disjoint datasets (with the same superclasses).
+        Get information about mapping between ImageNet wnids/class numbers/class names.
 
         Args:
-            level (int): Depth in hierarchy (from root node)
-            Nsubclasses (int): Number of subclasses per superclass
-                               (not used when balanced is True)
-            balanced (bool): Whether or not the dataset should be
-                             balanced over superclasses
-            split ("good"/"bad"/"rand"/None): If None, superclasses are
-                             not partitioned into two disjoint datasets.
-                             If not None, determines whether the subclass
-                             partitioning should be more or less 
-                             adversarial or random
-            rng (RandomState): Random number generator
+            ds_path (str) : Path to ImageNet dataset
+            ds_info_path (str) : Path to supplementary files for the ImageNet dataset 
+                                 ('wordnet.is_a.txt', 'words.txt', 'imagenet_class_index.json')
+                                 which can be obtained from http://image-net.org/download-API.
+
+        """
+        files = os.listdir(os.path.join(ds_path, 'train'))
+        in_wnids = [f for f in files if f[0]=='n'] 
+
+        f = open(os.path.join(ds_info_path, 'words.txt'))
+        wnid_to_name = [l.strip() for l in f.readlines()]
+        wnid_to_name = {l.split('\t')[0]: l.split('\t')[1] \
+                             for l in wnid_to_name}
+
+        with open(os.path.join(ds_info_path, 'imagenet_class_index.json'), 'r') as f:
+            base_map = json.load(f)
+            wnid_to_num = {v[0]: int(k) for k, v in base_map.items()}
+            num_to_name = {int(k): v[1] for k, v in base_map.items()}
+
+        return in_wnids, wnid_to_name, wnid_to_num, num_to_name
+
+    def get_node(self, wnid):
+        """
+        Add node to tree.
+
+        Args:
+            wnid (str) : WordNet ID for synset represented by node
 
         Returns:
-            subclass_ranges (list): Each entry is a list of subclasses 
-                                    for a given superclass in the dataset
-            label_map (dict): Map from class number to superclass name 
-            subclass_tuple (tuple): Tuple of lists capturing the split.
-                                    If split is None, lists are empty.
-                                    Otherwise, each list entry is 
-                                    a list of subclasses for a given 
-                                    superclass in the dataset
-            superclasses (list): WordNet IDs of superclasses
-            all_subclasses (list): List of all possible subclasses included in 
-                                    superclass
+            A node object representing the specified wnid.
+        """
+        if wnid not in self.tree:
+            self.tree[wnid] = Node(wnid, name=self.wnid_to_name[wnid])
+        return self.tree[wnid]
 
-        """ 
 
-        rng = np.random.RandomState(random_seed)
-        hierarchy = self.hierarchy
+    def is_ancestor(self, ancestor_wnid, child_wnid):
+        """
+        Check if a node is an ancestor of another.
 
-        # Identify superclasses at this level
-        rel_nodes = sorted(list(hierarchy.get_nodes_at_level(
-                                            level, ancestor=ancestor)))
-        if verbose: hierarchy.print_node_info(rel_nodes)
+        Args:
+            ancestor_wnid (str) : WordNet ID for synset represented by ancestor node
+            child_wnid (str) : WordNet ID for synset represented by child node
 
-        # Count number of subclasses
-        in_desc = []
-        for n in rel_nodes:
-            dcurr = self.hierarchy.leaves_reachable(n)
-            in_desc.append(sorted(list(dcurr)))
-        min_desc = np.min([len(d) for d in in_desc])
-        assert min_desc > 0
+        Returns:
+            A boolean variable indicating whether or not the node is an ancestor
+        """
+        return (child_wnid in self.tree[ancestor_wnid].descendants_all)
 
-        # Determine number of subclasses to include per superclass
-        if Nsubclasses is None:
-            if balanced:
-                Nsubclasses = min_desc
-                if Nsubclasses % 2 != 0: Nsubclasses = max(2, Nsubclasses - 1)
-            else:
-                Nsubclasses = 1 if split is None else 2
+    
+    def get_descendants(self, node_wnid, in_imagenet=False):
+        """
+        Get all descendants of a given node.
 
-        # Find superclasses that have sufficient subclasses
-        superclass_idx = [i for i in range(len(rel_nodes)) 
-                          if len(in_desc[i]) >= Nsubclasses]
-        superclasses, all_subclasses = [rel_nodes[i] for i in superclass_idx], \
-                                        [in_desc[i] for i in superclass_idx]
+        Args:
+            node_wnid (str) : WordNet ID for synset for node
+            in_imagenet (bool) : If True, only considers descendants among 
+                                ImageNet synsets, else considers all possible
+                                descendants in the WordNet hierarchy
 
-        # Superclass names
-        label_map = {}
-        for ri, r in enumerate(superclasses):
-            label_map[ri] = self.hierarchy.node_name(r)
-
-        subclass_ranges, subclass_tuple = [], ([], [])
-        
-        if split is None:
-
-            if balanced:
-                Ns = [Nsubclasses] * len(all_subclasses)
-            else:
-                Ns = [len(d) for d in all_subclasses]
-            wnids = [list(rng.choice(d, n, replace=False))
-                                   for n, d in zip(Ns, all_subclasses)] 
-            subclass_ranges = [[self.hierarchy.WORDNET_TO_NUM[w] for w in c] for c in wnids]
+        Returns:
+            A set of wnids corresponding to all the descendants
+        """        
+        if in_imagenet:
+            return set([self.wnid_to_num[ww] for ww in self.tree[node_wnid].descendants_all
+                        if ww in set(self.in_wnids)])
         else:
-            for sci, sc in enumerate(sorted(superclasses)):
-                class_tup = self.split_superclass(sc, Nsubclasses=Nsubclasses, 
-                                                     balanced=balanced,
-                                                     split_type=split, rng=rng)
-                subclass_tuple[0].append(class_tup[0])
-                subclass_tuple[1].append(class_tup[1])
-                subclass_ranges.append(class_tup[0] + class_tup[1])
+            return self.tree[node_wnid].descendants_all
+    
+    def get_superclasses(self, n_superclasses, 
+                         ancestor_wnid=None, superclass_lowest=None, 
+                         balanced=True):
+        """
+        Get superclasses by grouping together classes from the ImageNet dataset.
 
-        return subclass_ranges, label_map, subclass_tuple, superclasses, all_subclasses
+        Args:
+            n_superclasses (int) : Number of superclasses desired
+            ancestor_wnid (str) : (optional) WordNet ID that can be used to specify
+                                common ancestor for the selected superclasses
+            superclass_lowest (set of str) : (optional) Set of WordNet IDs of nodes
+                                that shouldn't be further sub-classes
+            balanced (bool) : If True, all the superclasses will have the same number
+                            of ImageNet subclasses
+
+        Returns:
+            superclass_wnid (list): List of WordNet IDs of superclasses
+            class_ranges (list of sets): List of ImageNet subclasses per superclass
+            label_map (dict): Mapping from class number to human-interpretable description
+                            for each superclass
+        """             
+        
+        assert superclass_lowest is None or \
+               not any([self.is_ancestor(s1, s2) for s1, s2 in product(superclass_lowest, superclass_lowest)])
+         
+        superclass_info = []
+        for (wnid, ndesc_in, ndesc_all) in self.wnid_sorted:
+            
+            if len(superclass_info) == n_superclasses:
+                break
+                
+            if ancestor_wnid is None or self.is_ancestor(ancestor_wnid, wnid):
+                keep_wnid = [True] * (len(superclass_info) + 1)
+                superclass_info.append((wnid, ndesc_in))
+                
+                for ii, (w, d) in enumerate(superclass_info):
+                    if self.is_ancestor(w, wnid):
+                        if superclass_lowest and w in superclass_lowest:
+                            keep_wnid[-1] = False
+                        else:
+                            keep_wnid[ii] = False
+                
+                for ii in range(len(superclass_info) - 1, -1, -1):
+                    if not keep_wnid[ii]:
+                        superclass_info.pop(ii)
+            
+        superclass_wnid = [w for w, _ in superclass_info]
+        class_ranges, label_map = self.get_subclasses(superclass_wnid, 
+                                    balanced=balanced)
+                
+        return superclass_wnid, class_ranges, label_map
 
 
-# Some standard datasets from the BREEDS paper.
+    def get_subclasses(self, superclass_wnid, balanced=True):
+        """
+        Get ImageNet subclasses for a given set of superclasses from the WordNet 
+        hierarchy. 
 
-def Entity13(info_dir, split=None):
+        Args:
+            superclass_wnid (list): List of WordNet IDs of superclasses
+            balanced (bool) : If True, all the superclasses will have the same number
+                            of ImageNet subclasses
+
+        Returns:
+            class_ranges (list of sets): List of ImageNet subclasses per superclass
+            label_map (dict): Mapping from class number to human-interpretable description
+                            for each superclass
+        """      
+        ndesc_min = min([self.tree[w].descendant_count_in for w in superclass_wnid]) 
+        class_ranges, label_map = [], {}
+        for ii, w in enumerate(superclass_wnid):
+            descendants = self.get_descendants(w, in_imagenet=True)
+            if balanced and len(descendants) > ndesc_min:
+                descendants = set([dd for ii, dd in enumerate(sorted(list(descendants))) if ii < ndesc_min])
+            class_ranges.append(descendants)
+            label_map[ii] = self.tree[w].name
+            
+        for i in range(len(class_ranges)):
+            for j in range(i + 1, len(class_ranges)):
+                assert(len(class_ranges[i].intersection(class_ranges[j])) == 0)
+                
+        return class_ranges, label_map
+
+def common_superclass_wnid(group_name):
     """
-    ENTITY-13 Dataset
-    Args:
-        info_dir (str) : Path to ImageNet information files
-        split ("good"/"bad"/"rand"/None): Nature of subclass
-    Returns:
-        subclass_ranges (list): Each entry is a list of subclasses 
-                                for a given superclass in the dataset
-        label_map (dict): Map from class number to superclass name 
-        subclass_tuple (tuple): Tuple of lists capturing the split; 
-                                each list entry is 
-                                a list of subclasses for a given 
-                                superclass in the dataset
-        superclasses (list): WordNet IDs of superclasses
-        all_subclasses (list): List of all possible subclasses included in 
-                                superclass
+        Get WordNet IDs of common superclasses. 
 
-    """ 
+        Args:
+            group_name (str): Name of group
 
-    DG = BreedsDatasetGenerator(info_dir)
-    return DG.get_superclasses(level=3, 
-                       ancestor=None,
-                       Nsubclasses=20, 
-                       split=split, 
-                       balanced=True, 
-                       random_seed=2,
-                       verbose=False)
+        Returns:
+            superclass_wnid (list): List of WordNet IDs of superclasses
+        """    
+    common_groups = {
 
-def Entity30(info_dir, split=None):
-    """
-    ENTITY-30 Dataset
-    Args:
-        info_dir (str) : Path to ImageNet information files
-        split ("good"/"bad"/"rand"/None): Nature of subclass
-    Returns:
-        subclass_ranges (list): Each entry is a list of subclasses 
-                                for a given superclass in the dataset
-        label_map (dict): Map from class number to superclass name 
-        subclass_tuple (tuple): Tuple of lists capturing the split; 
-                                each list entry is 
-                                a list of subclasses for a given 
-                                superclass in the dataset
-        superclasses (list): WordNet IDs of superclasses
-        all_subclasses (list): List of all possible subclasses included in 
-                                superclass
+        # ancestor_wnid = 'n00004258'
+        'living_9': ['n02084071', #dog, domestic dog, Canis familiaris
+                    'n01503061', # bird
+                    'n01767661', # arthropod
+                    'n01661091', # reptile, reptilian
+                    'n02469914', # primate
+                    'n02512053', # fish
+                    'n02120997', # feline, felid
+                    'n02401031', # bovid
+                    'n01627424', # amphibian
+                    ],
 
-    """ 
-    DG = BreedsDatasetGenerator(info_dir)
-    return DG.get_superclasses(level=4, 
-                       ancestor=None,
-                       Nsubclasses=8, 
-                       split=split, 
-                       balanced=True, 
-                       random_seed=2,
-                       verbose=False)
+        'mixed_10': [
+                     'n02084071', #dog,
+                     'n01503061', #bird 
+                     'n02159955', #insect 
+                     'n02484322', #monkey 
+                     'n02958343', #car 
+                     'n02120997', #feline 
+                     'n04490091', #truck 
+                     'n13134947', #fruit 
+                     'n12992868', #fungus 
+                     'n02858304', #boat 
+                     ],
 
-def Living17(info_dir, split=None):
-    """
-    LIVING-17 Dataset
-    Args:
-        info_dir (str) : Path to ImageNet information files
-        split ("good"/"bad"/"rand"/None): Nature of subclass
-    Returns:
-        subclass_ranges (list): Each entry is a list of subclasses 
-                                for a given superclass in the dataset
-        label_map (dict): Map from class number to superclass name 
-        subclass_tuple (tuple): Tuple of lists capturing the split; 
-                                each list entry is 
-                                a list of subclasses for a given 
-                                superclass in the dataset
-        superclasses (list): WordNet IDs of superclasses
-        all_subclasses (list): List of all possible subclasses included in 
-                                superclass
+        'mixed_13': ['n02084071', #dog,
+                     'n01503061', #bird (52)
+                     'n02159955', #insect (27)
+                     'n03405725', #furniture (21)
+                     'n02512053', #fish (16),
+                     'n02484322', #monkey (13)
+                     'n02958343', #car (10)
+                     'n02120997', #feline (8),
+                     'n04490091', #truck (7)
+                     'n13134947', #fruit (7)
+                     'n12992868', #fungus (7)
+                     'n02858304', #boat (6)  
+                     'n03082979', #computer(6)
+                    ],
 
-    """ 
-    DG = BreedsDatasetGenerator(info_dir)
-    return DG.get_superclasses(level=5, 
-                       ancestor="n00004258",
-                       Nsubclasses=4, 
-                       split=split, 
-                       balanced=True, 
-                       random_seed=2,
-                       verbose=False)
+        # Dataset from Geirhos et al., 2018: arXiv:1811.12231
+        'geirhos_16': ['n02686568', #aircraft (3)
+                       'n02131653', #bear (3)
+                       'n02834778', #bicycle (2)
+                       'n01503061', #bird (52)
+                       'n02858304', #boat (6)
+                       'n02876657', #bottle (7)
+                       'n02958343', #car (10)
+                       'n02121808', #cat (5)
+                       'n03001627', #char (4)
+                       'n03046257', #clock (3)
+                       'n02084071', #dog (116)
+                       'n02503517', #elephant (2)
+                       'n03614532', #keyboard (3)
+                       'n03623556', #knife (2)
+                       'n03862676', #oven (2)
+                       'n04490091', #truck (7)
+                      ],
+        'big_12':  ['n02084071', #dog (100+)
+                     'n04341686', #structure (55)
+                     'n01503061', #bird (52)
+                     'n03051540', #clothing (48)
+                     'n04576211', #wheeled vehicle
+                     'n01661091', #reptile, reptilian (36)
+                     'n02075296', #carnivore
+                     'n02159955', #insect (27)
+                     'n03800933', #musical instrument (26)
+                     'n07555863', #food (24)
+                     'n03405725', #furniture (21)
+                     'n02469914', #primate (20)
+                   ],
+        'mid_12':  ['n02084071', #dog (100+)
+                      'n01503061', #bird (52)
+                      'n04576211', #wheeled vehicle
+                      'n01661091', #reptile, reptilian (36)
+                      'n02075296', #carnivore
+                      'n02159955', #insect (27)
+                      'n03800933', #musical instrument (26)
+                      'n07555863', #food (24)
+                      'n03419014', #garment (24)
+                      'n03405725', #furniture (21)
+                      'n02469914', #primate (20)
+                      'n02512053', #fish (16)
+                    ]
+    }
 
-def Nonliving26(info_dir, split=None):
-    """
-    NONLIVING-26 Dataset
-    Args:
-        info_dir (str) : Path to ImageNet information files
-        split ("good"/"bad"/"rand"/None): Nature of subclass
-    Returns:
-        subclass_ranges (list): Each entry is a list of subclasses 
-                                for a given superclass in the dataset
-        label_map (dict): Map from class number to superclass name 
-        subclass_tuple (tuple): Tuple of lists capturing the split; 
-                                each list entry is 
-                                a list of subclasses for a given 
-                                superclass in the dataset
-        superclasses (list): WordNet IDs of superclasses
-        all_subclasses (list): List of all possible subclasses included in 
-                                superclass
-
-    """ 
-    DG = BreedsDatasetGenerator(info_dir)
-    return DG.get_superclasses(level=5, 
-                       ancestor="n00021939",
-                       Nsubclasses=4, 
-                       split=split, 
-                       balanced=True, 
-                       random_seed=2,
-                       verbose=False)
+    if group_name in common_groups:
+        superclass_wnid = common_groups[group_name]
+        return superclass_wnid
+    else:
+        raise ValueError("Custom group does not exist")
